Add selection sort implementation in C (#18)
* C implementation of Selection sort

This program prompts the user to input the length of an array, loads it with values, and sorts it using selection sort in both ascending and descending order. The maximum length of the array is currently set to 15, but can be changed. The implementation includes functions for loading the array, printing the array, and sorting the array in ascending and descending order.